Two big changes:
 - Added multithreading of API requests through the use of Coro
   and Coro::LWP. This allows the script to work much faster.

 - the characters and char_recipe tables have been split up into
   smaller per-realm tables. This makes both queries and inserts
   much faster.

Postgres's table inheritance/partitioning was considered, but
due to the lack of constraint inheritance, manual partitioning
was decided to be a better solution.

Since the database schema is now much more complex, the .sql
commands used to reset the database have been extended.

Also, various refactoring, cleaning up, etc. It's a big commit.